+++ test/test_bucket.cc
@@ -33,7 +33,7 @@ void MainPosttest() {
 
 void TestBlobCreates(hapi::Hermes *hermes) {
   auto bkt = hermes->GetBucket("hello");
-  int num_blobs = 256;
+  size_t num_blobs = 256;
   size_t blob_size = KILOBYTES(4);
   hermes::api::Context ctx;
   hermes::BlobId blob_id;
@@ -87,7 +87,7 @@ void TestBucketRename(hapi::Hermes *hermes) {
 
 void TestBucketClear(hapi::Hermes *hermes) {
   auto bkt = hermes->GetBucket("hello");
-  int num_blobs = 16;
+  size_t num_blobs = 16;
   size_t blob_size = MEGABYTES(32);
   hermes::api::Context ctx;
   hermes::BlobId blob_id;
@@ -109,7 +109,7 @@ void TestBucketClear(hapi::Hermes *hermes) {
 
 void TestBucketDestroy(hapi::Hermes *hermes) {
   auto bkt = hermes->GetBucket("hello");
-  int num_blobs = 1;
+  size_t num_blobs = 1;
   size_t blob_size = MEGABYTES(32);
   hermes::api::Context ctx;
   hermes::BlobId blob_id;
